% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ho2.R
\name{ho2_cereb}
\alias{ho2_cereb}
\title{Harvard-Oxford Atlas 2.0 Cerebellar}
\usage{
ho2_cereb()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cerebellar).
}
\description{
Cerebellar segmentation from the Harvard-Oxford Atlas 2.0
(HOA-2), including cerebellar cortex and white matter.
Contains 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}}.
}
\examples{
ho2_cereb()
plot(ho2_cereb())
}
\references{
Rushmore RJ, et al. (2022). Frontiers in
Neuroanatomy 16:1035420.
(\doi{10.3389/fnana.2022.1035420})
}
\seealso{
Other ggseg_atlases: 
\code{\link{ho2_cort}()},
\code{\link{ho2_sub}()},
\code{\link{ho_cort}()},
\code{\link{ho_sub}()}
}
\concept{cerebellar_atlases}
\concept{ggseg_atlases}
