% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegHO-package.R
\docType{package}
\name{ggsegHO-package}
\alias{ggsegHO}
\alias{ggsegHO-package}
\title{ggsegHO: Harvard-Oxford Atlas for the 'ggseg' Ecosystem}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Harvard-Oxford cortical, subcortical, and cerebellar atlases for the 'ggseg' ecosystem. Includes the original Harvard-Oxford atlas (FSL) and the Harvard-Oxford Atlas 2.0 (HOA-2, Rushmore et al. 2022) as 'ggseg_atlas' objects with 2D polygon and 3D mesh geometry.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegHO}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegHO/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Pineiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
