% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ho2.R
\name{ho2_cort}
\alias{ho2_cort}
\title{Harvard-Oxford Atlas 2.0 Cortical}
\usage{
ho2_cort()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Cortical parcellation with 49 regions per hemisphere from the
Harvard-Oxford Atlas 2.0 (HOA-2), based on 50 HCP subjects.
Contains 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and
3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
ho2_cort()
plot(ho2_cort())
}
\references{
Rushmore RJ, et al. (2022). Frontiers in
Neuroanatomy 16:1035420.
(\doi{10.3389/fnana.2022.1035420})
}
\seealso{
Other ggseg_atlases: 
\code{\link{ho2_cereb}()},
\code{\link{ho2_sub}()},
\code{\link{ho_cort}()},
\code{\link{ho_sub}()}

Other cortical_atlases: 
\code{\link{ho_cort}()}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
