% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ho2.R
\name{ho2_sub}
\alias{ho2_sub}
\title{Harvard-Oxford Atlas 2.0 Subcortical}
\usage{
ho2_sub()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Subcortical segmentation with 19 structures from the
Harvard-Oxford Atlas 2.0 (HOA-2), based on 100 HCP subjects.
Contains 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and
3D mesh data for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
ho2_sub()
plot(ho2_sub())
}
\references{
Rushmore RJ, et al. (2022). Frontiers in
Neuroanatomy 16:1035420.
(\doi{10.3389/fnana.2022.1035420})
}
\seealso{
Other ggseg_atlases: 
\code{\link{ho2_cereb}()},
\code{\link{ho2_cort}()},
\code{\link{hoCort}()},
\code{\link{hoSub}()}

Other subcortical_atlases: 
\code{\link{hoSub}()}
}
\concept{ggseg_atlases}
\concept{subcortical_atlases}
