% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ho.R
\name{ho_sub}
\alias{ho_sub}
\title{Harvard-Oxford Subcortical Atlas}
\usage{
ho_sub()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Subcortical segmentation from the Harvard-Oxford atlas
distributed with FSL, drawn on a grey-brain anatomical context.
Contains 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and
3D mesh data for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
ho_sub()
plot(ho_sub())
}
\references{
Makris, et al. (2006) Schizophrenia research
83(2-3):155-151
(\doi{10.1016/j.schres.2005.11.020})
}
\seealso{
Other ggseg_atlases: 
\code{\link{ho2_cereb}()},
\code{\link{ho2_cort}()},
\code{\link{ho2_sub}()},
\code{\link{ho_cort}()}

Other subcortical_atlases: 
\code{\link{ho2_sub}()}
}
\concept{ggseg_atlases}
\concept{subcortical_atlases}
