% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hoCort}
\alias{hoCort}
\title{Harvard-Oxford Cortical Atlas}
\usage{
hoCort()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Cortical parcellation with 48 regions per hemisphere from the
Harvard-Oxford atlas distributed with FSL. Contains 2D polygon
geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and 3D vertex indices for
\code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
hoCort()
}
\references{
Makris, et al. (2006) Schizophrenia research 83(2-3):155-151
(\doi{10.1016/j.schres.2005.11.020})
}
\seealso{
Other ggseg_atlases: 
\code{\link{hoSub}()}
}
\concept{ggseg_atlases}
