% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hoCort}
\alias{hoCort}
\title{Harvard-Oxford Cortical Atlas}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
data(hoCort)
}
\description{
Brain atlas for the Harvard-Oxford cortical parcellation with
48 regions per hemisphere. Contains 2D polygon geometry for
\code{\link[ggseg:ggbrain]{ggseg::geom_brain()}}.
}
\examples{
data(hoCort)
hoCort
}
\references{
Makris, et al. (2006) Schizophrenia research 83(2-3):155-151
(\doi{10.1016/j.schres.2005.11.020})
}
\concept{ggseg_atlases}
\keyword{datasets}
